Refactor Playground components: rename and reorganize imports for clarity and consistency
