+++ client/src/layouts/PlaygroundLayout/PlaygroundLayout.jsx
@@ -3,15 +3,15 @@ import "./PlaygroundLayout.css";
 
 // Import components
 import { NavBar } from "../../components/NavBar/NavBar";
-import { Header } from "../../components/Header/Header";
-import { PlaygroundChat } from "../../components/PlaygroundChat/PlaygroundChat";
+import { PlaygroundHeader } from "../../components/Playground/PlaygroundHeader/PlaygroundHeader";
+import { PlaygroundChat } from "../../components/Playground/PlaygroundChat/PlaygroundChat";
 
 export const PlaygroundLayout = ({ children }) => {
   return (
     <div>
       <NavBar />
       <main id="playground-layout">
-        <Header />
+        <PlaygroundHeader />
         <div className="playground-content">
           <section style={{ padding: "15px" }}>{children}</section>
           <PlaygroundChat />
